feat: create error card and spearate functionality hook for the home page
