+++ weather-app/src/components/ThemeSwitcher.tsx
@@ -7,7 +7,7 @@ export const ThemeSwitcher = () => {
 
   return (
     <button
-      className=" absolute size-20 bottom-8 right-0 p-2 pr-6 border-2 border-blue-950 border-r-0 rounded-l-2xl"
+      className=" fixed size-20 bottom-8 right-0 p-2 pr-6 border-2 border-blue-950 border-r-0 rounded-l-2xl"
       onClick={toggleDarkMode}
     >
       {darkMode ? <LightModeSVG /> : <DarkModeSVG />}
