+++ weather-app/src/components/cards/InfoCard.tsx
