+++ weather-app/src/pages/home/Home.tsx
@@ -1,44 +1,41 @@
-import type { FormEvent } from "react";
-import { useLazyGetCurrentWeatherByCityQuery } from "../../services/weatherApi";
-
-import { useAppDispatch, useAppSelector } from "../../typedHooks";
-
-import { reset, update } from "./searchSlice";
-import { InfoCard } from "../../components/InfoCard";
+import { InfoCard } from "../../components/cards/InfoCard";
+import { ErrorCard } from "../../components/cards/ErrorCard";
+import { useWeatherSearch } from "./useWeatherSearch";
 
 export const Home = () => {
-  const searchValue = useAppSelector((state) => state.search.value);
-  const dispatch = useAppDispatch();
-  const [trigger, { data, isLoading, isError }] =
-    useLazyGetCurrentWeatherByCityQuery();
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    await trigger(searchValue);
-    console.log(data);
-    dispatch(reset());
-  };
+  const {
+    searchValue,
+    handleSubmit,
+    isError,
+    isLoading,
+    data,
+    handleSearchChange,
+  } = useWeatherSearch();
 
   return (
     <div className=" flex flex-col items-center gap-12">
       <div className=" flex flex-col items-center gap-24">
         <h1 className=" text-center text-4xl">Welcome to Weather App</h1>
-        <form onSubmit={async (e) => await handleSubmit(e)}>
+        <form onSubmit={handleSubmit}>
           <label className=" flex flex-col text-xl">
             Checkout any city's current weather
             <input
               value={searchValue}
               minLength={2}
               required
-              onChange={(e) => dispatch(update(e.target.value))}
+              onChange={handleSearchChange}
               className="border-2  rounded-4xl px-1 py-0.5"
               type="search"
             />
           </label>
-          <button type="submit">Search</button>
+          <button type="submit" disabled={isLoading}>
+            {isLoading ? "Loading..." : "Search"}
+          </button>
         </form>
       </div>
+      {isError && <ErrorCard />}
       {isLoading && <p>Loading ...</p>}
-      {data && <InfoCard data={data} />}
+      {data && !isError && <InfoCard data={data} />}
     </div>
   );
 };
